Update version history popover
